+++ tool/tsh/common/tsh.go
@@ -1339,8 +1339,6 @@ func Run(ctx context.Context, args []string, opts ...CliOption) error {
 		puttyConfig.Hidden()
 	}
 
-	printDocs := app.Command("docs", "Print CLI documentation").Hidden()
-
 	// FIDO2, TouchID and WebAuthnWin commands.
 	f2 := fido2.NewCommand(app)
 	tid := touchid.NewCommand(app)
